requirements.txt
python-dotenv
openai
pyyaml
google-generativeai